Add files via upload
Given a project to develop a database that would analyze and visualize hotel booking data, I created visuals that answered the questions: 1. Is the hotel revenue growing by year? 2. Should the parking lot size be increased? 3. What trends  in average daily rate and guests can be seen in the data?